added description div in app overview and job overview page
